created announcement with auth meaning it can only be seen by the account user

+++ CCIS-MIRROR/database/migrations/2026_03_11_002658_create_table_announcement_attachment.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('table_announcement_attachment', function (Blueprint $table) {
             $table->integer('attachment_id')->primary();
-            $table->integer('announcement_id');
+            $table->unsignedBigInteger('announcement_id');
             $table->string('file_path');
             $table->string('file_type');
             //
